requirements.txt
mss
Pillow
pywin32
pystray
keyboard
